+++ src/client/PteroClient.js
@@ -1,9 +1,9 @@
 const { EventEmitter } = require('events');
-const ClientRequestManager = require('./managers/ClientRequestManager');
-const ClientServerManager = require('./managers/ClientServerManager');
+const ClientRequestManager = require('./ClientRequestManager');
+const ClientServerManager = require('./ClientServerManager');
 const { ClientUser } = require('../structures/User');
-const WebSocketManager = require('./managers/WebSocketManager');
-const endpoints = require('./managers/endpoints');
+const WebSocketManager = require('./WebSocketManager');
+const endpoints = require('./endpoints');
 const presets = require('../structures/Presets');
 
 /**
